Add files via upload
Although we would want to train more different datasets, the network asks strictly that the input vector should be N*1*28*28. Since most datasets have different sizes of data, we could not move forward to training more networks. Xilinx is developing a library called Brevitas which could be used for more custom datasets but since it’s still under development and there are still errors when people train their custom dataset, we could not make good use of it in such a limited time. So, in the future, if Brevitas is well established, we may be able to train networks for more different datasets.